+++ src/main/java/com/webank/oracle/event/callback/AbstractEventCallback.java
@@ -69,11 +69,12 @@ public abstract class AbstractEventCallback extends EventLogPushWithDecodeCallba
 
     /**
      * 部署合约
-     *
      * @param chainId
      * @param group
+     * @param coreContractAddress
+     * @param vrfContractAddress
      */
-    public abstract String deployContract(int chainId, int group);
+    public abstract String deployOrLoadContract(int chainId, int group, String coreContractAddress, String vrfContractAddress);
 
     /**
      * 解析日志
@@ -181,8 +182,9 @@ public abstract class AbstractEventCallback extends EventLogPushWithDecodeCallba
         // set decoder
         TransactionDecoder decoder = new TransactionDecoder(abi);
         setDecoder(decoder);
-        // TODO. deploy????
-        String contractAddress = this.deployContract(eventRegister.getChainId(), eventRegister.getGroup());
+        // TODO. deploy and load
+
+        String contractAddress = this.deployOrLoadContract(eventRegister.getChainId(), eventRegister.getGroup(),eventRegister.getOracleCoreContractAddress(),eventRegister.getVrfContractAddress());
         if (StringUtils.isBlank(contractAddress)) {
             log.error("Deploy contract error");
             return;
