+++ src/main/resources/application.yml
@@ -45,7 +45,9 @@ event:
   eventRegisters:
     # list of EventRegister objects. Use default value when property not configured.
     # Oracle contracts will be auto deployed when contractAddress is blank
-    - {chainId: 1, group: 1, oracleCoreContractAddress: ""}
+    - {chainId: 1, group: 1}
+   # - {chainId: 1, group: 2}
+
 
 #log config
 logging:
